Bug fix
When the clock almost stops, there is lag updating the clock face.
Instead of resetting the clock when timeLeft <= 0, just set the timeLeft
to 0.

Update: Clock face working, but the css is messy. Although it is fixed
(sth wrong with .display), I still do not fully understand it.

Meanwhile, we only have one clock dom object and both clock wants it. so
it is better to just hide one of it.